buildpkg: make repository name optarg; add sig check deploypkg: add some checks for existing pkg & sig

+++ lib/util-pkg-publish.sh
@@ -12,7 +12,10 @@
 repo_update(){
     local repo="$1" arch="$2" pkg="$3" action="$4" cmd
     if [[ $action == "add" ]];then
-        ln -s ${cache_dir_pkg}/$arch/$pkg{,.sig} ${repos_local}/$repo/os/$arch/
+        if [[ -f ${repos_local}/$repo/os/$arch/$pkg{,.sig} ]];then
+            rm -v ${repos_local}/$repo/os/$arch/$pkg{,.sig}
+            ln -s ${cache_dir_pkg}/$arch/$pkg{,.sig} ${repos_local}/$repo/os/$arch/
+        fi
     fi
     repo-$action -R ${repos_local}/$repo/os/$arch/$repo.db.tar.xz ${repos_local}/$repo/os/$arch/$pkg
 }
